+++ packages/@orchidui-vue/src/StoreDesign/SDSidebar/SDSidebar.sample.js
@@ -1,4 +1,3 @@
-
 const SDSidebarSample = [
   {
     type: "list",
@@ -11,11 +10,11 @@ const SDSidebarSample = [
       },
       {
         name: "sections",
-        label: "Sections"
+        label: "Sections",
       },
       {
         name: "footer",
-        label: "Footer"
+        label: "Footer",
       },
     ],
   },
@@ -58,37 +57,4 @@ const SDSidebarSample = [
   },
 ];
 
-
-const RequestFormStyles = [
-  {
-    type: "title",
-    label: "Choose Font",
-  },
-  {
-    type: "font",
-    name: "font_heading",
-    label: "Heading",
-  },
-  {
-    type: "colors",
-    name: "colors",
-    label: "Choose Colors",
-  },
-  {
-    type: "range",
-    name: "rounded",
-    range: [0, 100],
-    label: "Card appearance",
-    hint: "Rounded corner",
-  },
-  {
-    type: "shadow",
-    name: "shadow",
-    label: "Shadow Styles",
-  },
-];
-
-export {
-  SDSidebarSample,
-  RequestFormStyles,
-};
+export { SDSidebarSample };
